+++ arch/x86/kernel/cpu/resctrl/monitor.c
@@ -16,11 +16,11 @@
  */
 
 #include <linux/module.h>
+#include <linux/resctrl.h>
 #include <linux/sizes.h>
 #include <linux/slab.h>
 
 #include <asm/cpu_device_id.h>
-#include <asm/resctrl.h>
 
 #include "internal.h"
